Improve migration tools
